requirements.txt
smolagents
sqlalchemy
gradio
requests
python-dotenv
groq
litellm